Refactor filter dropdown
