+++ src/Homes/Filters/Guests.js
@@ -1,26 +1,11 @@
 import React from "react";
 import styled from "styled-components";
-import MediaQuery from "react-responsive";
 
 import bp from "../../breakpoints";
 import NumericInput from "../../UI/NumericInput";
-import Dropdown from "../../UI/Dropdown";
-import {
-  Container,
-  MobileHeader,
-  CloseButton,
-  ResetButton,
-  ControlsGroup,
-  Label,
-  Caption,
-  Footer,
-  CancelButton,
-  ApplyButton,
-  MobileFooter,
-  MobilePrimaryButton
-} from "./styled";
+import Filter from "./Filter";
 
-import closeSvg from "./close.svg";
+import { ControlsGroup, Label, Caption } from "./styled";
 
 export const Body = styled.div`
   padding: 40px 8px 8px;
@@ -50,67 +35,51 @@ export default class extends React.Component {
     if (guestsTotal > 1) buttonText = `${guestsTotal} guests`;
 
     return (
-      <Dropdown buttonText={buttonText}>
-        <Container>
-          <MediaQuery maxDeviceWidth={bp.sm - 1}>
-            <MobileHeader>
-              <CloseButton>
-                <img src={closeSvg} alt="" />
-              </CloseButton>
-              Guests
-              <ResetButton onClick={this.reset}>Reset</ResetButton>
-            </MobileHeader>
-          </MediaQuery>
-          <Body>
-            <ControlsGroup>
-              <div>
-                <Label>Adults</Label>
-              </div>
-              <NumericInput
-                min={1}
-                max={this.props.maxGuests - this.state.children}
-                value={this.state.adults}
-                onValueChange={value => this.setState({ adults: value })}
-              />
-            </ControlsGroup>
-            <ControlsGroup>
-              <div>
-                <Label>Children</Label>
-                <Caption>Ages 2 &ndash; 12</Caption>
-              </div>
-              <NumericInput
-                min={0}
-                max={this.props.maxGuests - this.state.adults}
-                value={this.state.children}
-                onValueChange={value => this.setState({ children: value })}
-              />
-            </ControlsGroup>
-            <ControlsGroup>
-              <div>
-                <Label>Infants</Label>
-                <Caption>Under 2</Caption>
-              </div>
-              <NumericInput
-                min={0}
-                max={this.props.maxInfants}
-                value={this.state.infants}
-                onValueChange={value => this.setState({ infants: value })}
-              />
-            </ControlsGroup>
-          </Body>
-          <MediaQuery minDeviceWidth={bp.sm}>
-            <Footer>
-              <CancelButton>Cancel</CancelButton>
-              <ApplyButton>Apply</ApplyButton>
-            </Footer>
-          </MediaQuery>
-          <MediaQuery maxDeviceWidth={bp.sm - 1}>
-            <MobileFooter>
-              <MobilePrimaryButton>Save</MobilePrimaryButton>
-            </MobileFooter>
-          </MediaQuery>
-        </Container>
-      </Dropdown>
+      <Filter
+        buttonText={buttonText}
+        heading="Guests"
+        hasMobileHeaderSeparator
+        hasMobileFooter
+        onReset={this.reset}
+      >
+        <Body>
+          <ControlsGroup>
+            <div>
+              <Label>Adults</Label>
+            </div>
+            <NumericInput
+              min={1}
+              max={this.props.maxGuests - this.state.children}
+              value={this.state.adults}
+              onValueChange={value => this.setState({ adults: value })}
+            />
+          </ControlsGroup>
+          <ControlsGroup>
+            <div>
+              <Label>Children</Label>
+              <Caption>Ages 2 &ndash; 12</Caption>
+            </div>
+            <NumericInput
+              min={0}
+              max={this.props.maxGuests - this.state.adults}
+              value={this.state.children}
+              onValueChange={value => this.setState({ children: value })}
+            />
+          </ControlsGroup>
+          <ControlsGroup>
+            <div>
+              <Label>Infants</Label>
+              <Caption>Under 2</Caption>
+            </div>
+            <NumericInput
+              min={0}
+              max={this.props.maxInfants}
+              value={this.state.infants}
+              onValueChange={value => this.setState({ infants: value })}
+            />
+          </ControlsGroup>
+        </Body>
+      </Filter>
     );
   }
 }
